+++ backend/canisters/user_index/impl/src/lib.rs
@@ -8,13 +8,14 @@ use canister_state_macros::canister_state;
 use canister_timer_jobs::TimerJobs;
 use local_user_index_canister::Event as LocalUserIndexEvent;
 use model::local_user_index_map::LocalUserIndexMap;
+use model::pending_payments_queue::{BackdatedReferralReward, PendingPayment, PendingPaymentReason, PendingPaymentsQueue};
+use queries::referral_metrics::ReferralData;
 use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
 use types::{CanisterId, CanisterWasm, Cryptocurrency, Cycles, Milliseconds, TimestampMillis, Timestamped, UserId, Version};
 use utils::canister::{CanistersRequiringUpgrade, FailedUpgradeCount};
 use utils::canister_event_sync_queue::CanisterEventSyncQueue;
-use utils::consts::{SNS_GOVERNANCE_CANISTER_ID, SNS_LEDGER_CANISTER_ID};
 use utils::env::Environment;
 use utils::time::DAY_IN_MS;
 
@@ -112,6 +113,60 @@ impl RuntimeState {
         jobs::sync_events_to_local_user_index_canisters::start_job_if_required(self);
     }
 
+    pub fn queue_payment(&mut self, pending_payment: PendingPayment) {
+        self.data.pending_payments_queue.push(pending_payment);
+        jobs::make_pending_payments::start_job_if_required(self);
+    }
+
+    pub fn queue_backdated_referral_payments(&mut self) {
+        if self.data.backdated_referral_payments_added {
+            return;
+        }
+
+        let mut user_referrals_map: HashMap<UserId, ReferralData> = HashMap::new();
+        let now = self.env.now();
+
+        for user in self.data.users.iter() {
+            if let Some(referred_by) = user.referred_by {
+                if let Some(referrer) = self.data.users.get_by_user_id(&referred_by) {
+                    if referrer.diamond_membership_details.is_active(now) {
+                        let data = user_referrals_map.entry(referred_by).or_default();
+                        let icp_raised_for_paid_diamond: u64 =
+                            user.diamond_membership_details.payments().iter().map(|p| p.amount_e8s).sum();
+                        if icp_raised_for_paid_diamond > 0 {
+                            data.paid_diamond += 1;
+                            data.icp_raised_for_paid_diamond_e8s += icp_raised_for_paid_diamond;
+                        } else if user.diamond_membership_details.is_active(now) {
+                            data.unpaid_diamond += 1;
+                        } else {
+                            data.other += 1;
+                        }
+                    }
+                }
+            }
+        }
+
+        for (user_id, data) in user_referrals_map {
+            if data.paid_diamond > 0 || data.unpaid_diamond > 0 {
+                let amount: u64 = (data.icp_raised_for_paid_diamond_e8s / 2) + (data.unpaid_diamond as u64 * 75_000_000);
+                self.data.pending_payments_queue.push(PendingPayment {
+                    amount,
+                    currency: Cryptocurrency::InternetComputer,
+                    timestamp: self.env.now_nanos(),
+                    recipient: user_id.into(),
+                    reason: PendingPaymentReason::BackdatedReferralReward(BackdatedReferralReward {
+                        referrals_to_paid_members: data.paid_diamond,
+                        referrals_to_gifted_members: data.unpaid_diamond,
+                    }),
+                })
+            }
+        }
+
+        self.data.backdated_referral_payments_added = true;
+
+        jobs::make_pending_payments::start_job_if_required(self);
+    }
+
     pub fn metrics(&self) -> Metrics {
         let now = self.env.now();
         let canister_upgrades_metrics = self.data.canisters_requiring_upgrade.metrics();
@@ -165,6 +220,8 @@ struct Data {
     pub storage_index_user_sync_queue: OpenStorageUserSyncQueue,
     pub user_index_event_sync_queue: CanisterEventSyncQueue<LocalUserIndexEvent>,
     pub user_principal_migration_queue: UserPrincipalMigrationQueue,
+    #[serde(default)]
+    pub pending_payments_queue: PendingPaymentsQueue,
     pub platform_moderators: HashSet<UserId>,
     pub platform_operators: HashSet<UserId>,
     pub test_mode: bool,
@@ -173,9 +230,9 @@ struct Data {
     pub local_index_map: LocalUserIndexMap,
     pub timer_jobs: TimerJobs<TimerJob>,
     pub neuron_controllers_for_initial_airdrop: HashMap<UserId, Principal>,
-    pub openchat_governance_canister_id: CanisterId,
-    pub openchat_ledger_canister_id: CanisterId,
     pub internet_identity_canister_id: CanisterId,
+    #[serde(default)]
+    pub backdated_referral_payments_added: bool,
 }
 
 impl Data {
@@ -207,6 +264,7 @@ impl Data {
             storage_index_user_sync_queue: OpenStorageUserSyncQueue::default(),
             user_index_event_sync_queue: CanisterEventSyncQueue::default(),
             user_principal_migration_queue: UserPrincipalMigrationQueue::default(),
+            pending_payments_queue: PendingPaymentsQueue::default(),
             platform_moderators: HashSet::new(),
             platform_operators: HashSet::new(),
             test_mode,
@@ -215,9 +273,8 @@ impl Data {
             local_index_map: LocalUserIndexMap::default(),
             timer_jobs: TimerJobs::default(),
             neuron_controllers_for_initial_airdrop: HashMap::new(),
-            openchat_governance_canister_id: SNS_GOVERNANCE_CANISTER_ID,
-            openchat_ledger_canister_id: SNS_LEDGER_CANISTER_ID,
             internet_identity_canister_id,
+            backdated_referral_payments_added: false,
         };
 
         // Register the ProposalsBot
@@ -253,6 +310,7 @@ impl Default for Data {
             storage_index_user_sync_queue: OpenStorageUserSyncQueue::default(),
             user_index_event_sync_queue: CanisterEventSyncQueue::default(),
             user_principal_migration_queue: UserPrincipalMigrationQueue::default(),
+            pending_payments_queue: PendingPaymentsQueue::default(),
             platform_moderators: HashSet::new(),
             platform_operators: HashSet::new(),
             test_mode: true,
@@ -261,9 +319,8 @@ impl Default for Data {
             local_index_map: LocalUserIndexMap::default(),
             timer_jobs: TimerJobs::default(),
             neuron_controllers_for_initial_airdrop: HashMap::new(),
-            openchat_governance_canister_id: SNS_GOVERNANCE_CANISTER_ID,
-            openchat_ledger_canister_id: SNS_LEDGER_CANISTER_ID,
             internet_identity_canister_id: Principal::anonymous(),
+            backdated_referral_payments_added: false,
         }
     }
 }